Btn2 -> Btn으로 이름 변경
