+++ src/App.vue
@@ -1,24 +1,24 @@
 <template>
-  <!-- <SideBar></SideBar>
+  <SideBar></SideBar>
   <div id="main-wrapper">
     <Main></Main>
     <Basket></Basket>
     <Footer></Footer>
-  </div> -->
+  </div>
   <!-- <Join></Join> -->
   <!-- <SideBar></SideBar>
   <div id="main-wrapper">
     <history1></history1>
     <Footer></Footer>
-  </div> -->
-  <Login></Login>
+  </div>
+  <Login></Login> -->
 </template>
 
 <script lang="ts">
 import { Options, Vue } from 'vue-class-component';
 import HelloWorld from './components/HelloWorld.vue';
 import SideBar from './views/common/main/sidebar/Sidebar.vue';
-import Buttons from './views/common/components/Btn2.vue';
+import Buttons from './views/common/components/Btn.vue';
 import Join from './views/login/Join.vue';
 import modal1 from './views/common/components/Modal1.vue';
 import history1 from './views/purchase/purchase-history/PurchaseHistory.vue';
